Deviation from the Mean

Calculates the deviation from the mean.

d(x)

Arguments

x

Numeric matrix, data frame, or vector. Data matrix.

Value

A matrix.

Details

The \(n \times 1\) vector of deviations from the mean is given by

$$ \mathbf{d} = \mathbf{x} - \bar{x} $$

where \(\mathbf{x}\) is an \(n \times 1\) column vector and \(\bar{x}\) is the mean of \(\mathbf{x}\).

The \(n \times k\) matrix of deviations from the mean is given by

$$ \mathbf{D} = \mathbf{X} - \mathbf{1}_{n} \bar{\mathbf{x}} $$

where \(\mathbf{X}\) is an \(n \times 1\) matrix, \(\mathbf{1}_{n}\) is an \(n \times 1\) column vector of ones, and \(\bar{\mathbf{x}}\) is the \(k \times 1\) mean vector of \(\mathbf{X}\).

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

x <- matrix(
  rnorm(n = 5 * 5, mean = 100, sd = 15),
  ncol = 5
)

d(x)
#>            [,1]         [,2]       [,3]       [,4]       [,5]
#> [1,]   2.069034  -1.32860257  11.723138   7.274418 -28.281290
#> [2,]   2.370091   8.65254940  -9.763170   6.481931   1.823646
#> [3,]  25.984858 -14.25647566  12.102561 -13.433134  33.763979
#> [4,] -22.196261  -0.01088183  -1.287675  11.256817 -33.827187
#> [5,]  -8.227722   6.94341066 -12.774855 -11.580032  26.520852

x <- rnorm(5)

d(x)
#>            [,1]
#> [1,] -0.9585359
#> [2,] -0.6796615
#> [3,] -0.5643325
#> [4,]  1.7357666
#> [5,]  0.4667633